+++ src/Firebase/Factory.php
@@ -43,7 +43,6 @@ use Kreait\Firebase\Exception\MessagingApiExceptionConverter;
 use Kreait\Firebase\Exception\RuntimeException;
 use Kreait\Firebase\Http\HttpClientOptions;
 use Kreait\Firebase\Http\Middleware;
-use Kreait\Firebase\Project\ProjectId;
 use Kreait\Firebase\Value\Email;
 use Psr\Cache\CacheItemPoolInterface;
 use Psr\Http\Message\UriInterface;
@@ -73,7 +72,7 @@ class Factory
 
     protected ?CredentialsLoader $googleAuthTokenCredentials = null;
 
-    protected ?ProjectId $projectId = null;
+    protected ?string $projectId = null;
 
     protected ?string $clientEmail = null;
 
@@ -129,7 +128,7 @@ class Factory
     public function withProjectId(string $projectId): self
     {
         $factory = clone $this;
-        $factory->projectId = ProjectId::fromString($projectId);
+        $factory->projectId = $projectId;
 
         return $factory;
     }
@@ -282,7 +281,7 @@ class Factory
         return null;
     }
 
-    protected function getProjectId(): ProjectId
+    protected function getProjectId(): string
     {
         if ($this->projectId !== null) {
             return $this->projectId;
@@ -291,7 +290,7 @@ class Factory
         $serviceAccount = $this->getServiceAccount();
 
         if ($serviceAccount !== null) {
-            return $this->projectId = ProjectId::fromString($serviceAccount->getProjectId());
+            return $this->projectId = $serviceAccount->getProjectId();
         }
 
         if ($this->discoveryIsDisabled) {
@@ -303,11 +302,11 @@ class Factory
             && ($credentials instanceof ProjectIdProviderInterface)
             && ($projectId = $credentials->getProjectId())
         ) {
-            return $this->projectId = ProjectId::fromString($projectId);
+            return $this->projectId = $projectId;
         }
 
         if ($projectId = Util::getenv('GOOGLE_CLOUD_PROJECT')) {
-            return $this->projectId = ProjectId::fromString($projectId);
+            return $this->projectId = $projectId;
         }
 
         throw new RuntimeException('Unable to determine the Firebase Project ID');
@@ -347,7 +346,7 @@ class Factory
     protected function getDatabaseUri(): UriInterface
     {
         if ($this->databaseUri === null) {
-            $this->databaseUri = GuzzleUtils::uriFor(\sprintf(self::$databaseUriPattern, $this->getProjectId()->sanitizedValue()));
+            $this->databaseUri = GuzzleUtils::uriFor(\sprintf(self::$databaseUriPattern, $this->getProjectId()));
         }
 
         return $this->databaseUri;
@@ -356,7 +355,7 @@ class Factory
     protected function getStorageBucketName(): string
     {
         if ($this->defaultStorageBucket === null) {
-            $this->defaultStorageBucket = \sprintf(self::$storageBucketNamePattern, $this->getProjectId()->sanitizedValue());
+            $this->defaultStorageBucket = \sprintf(self::$storageBucketNamePattern, $this->getProjectId());
         }
 
         return $this->defaultStorageBucket;
@@ -406,7 +405,7 @@ class Factory
     {
         $keyStore = new HttpKeyStore(new Client(), $this->verifierCache);
 
-        $baseVerifier = new LegacyIdTokenVerifier($this->getProjectId()->sanitizedValue(), $keyStore);
+        $baseVerifier = new LegacyIdTokenVerifier($this->getProjectId(), $keyStore);
 
         if ($this->tenantId !== null) {
             $baseVerifier = new TenantAwareVerifier($this->tenantId->toString(), $baseVerifier);
@@ -433,7 +432,7 @@ class Factory
     public function createRemoteConfig(): Contract\RemoteConfig
     {
         $http = $this->createApiClient([
-            'base_uri' => "https://firebaseremoteconfig.googleapis.com/v1/projects/{$this->getProjectId()->value()}/remoteConfig",
+            'base_uri' => "https://firebaseremoteconfig.googleapis.com/v1/projects/{$this->getProjectId()}/remoteConfig",
         ]);
 
         return new RemoteConfig(new RemoteConfig\ApiClient($http));
@@ -447,7 +446,7 @@ class Factory
 
         $messagingApiClient = new Messaging\ApiClient(
             $this->createApiClient([
-                'base_uri' => 'https://fcm.googleapis.com/v1/projects/'.$projectId->value(),
+                'base_uri' => 'https://fcm.googleapis.com/v1/projects/'.$projectId,
             ]),
             $errorHandler
         );
@@ -478,18 +477,12 @@ class Factory
 
     public function createFirestore(): Contract\Firestore
     {
-        $config = [];
-        $projectId = $this->getProjectId();
-        $serviceAccount = $this->getServiceAccount();
+        $config = [
+            'projectId' => $this->getProjectId(),
+        ];
 
-        if ($serviceAccount !== null) {
+        if (($serviceAccount = $this->getServiceAccount()) !== null) {
             $config['keyFile'] = $serviceAccount->asArray();
-        } elseif ($this->discoveryIsDisabled) {
-            throw new RuntimeException('Unable to create a Firestore Client without credentials');
-        }
-
-        if ($projectId !== null) {
-            $config['projectId'] = $projectId->value();
         }
 
         try {
@@ -503,17 +496,13 @@ class Factory
 
     public function createStorage(): Contract\Storage
     {
-        $config = [];
-        $projectId = $this->getProjectId();
-        $serviceAccount = $this->getServiceAccount();
+        $config = [
+            'projectId' => $this->getProjectId(),
+        ];
 
-        if ($serviceAccount !== null) {
+        if (($serviceAccount = $this->getServiceAccount()) !== null) {
             $config['keyFile'] = $serviceAccount->asArray();
-        } elseif ($this->discoveryIsDisabled) {
-            throw new RuntimeException('Unable to create a Storage Client without credentials');
         }
-
-        $config['projectId'] = $projectId->value();
 
         try {
             $storageClient = new StorageClient($config);
@@ -541,7 +530,7 @@ class Factory
     public function getDebugInfo(): array
     {
         try {
-            $projectId = $this->getProjectId()->value();
+            $projectId = $this->getProjectId();
         } catch (Throwable $e) {
             $projectId = $e->getMessage();
         }